add response to done button

+++ RealmIOS/Select/SelectViewController.swift
@@ -64,14 +64,18 @@ class SelectViewController: UIViewController {
     }
     
     @IBAction func didTapRemoveAllActionButton(_ sender: Any) {
-        let alertController = UIAlertController(title: "Warning", message: "Delete All Data", preferredStyle: .alert)
-        let alertAction = UIAlertAction(title: "Ok", style: .default) { (_) in
-            self.realmService.deleteData()
+        if realmService.getUser().count > 0 {
+            let alertController = UIAlertController(title: "Warning", message: "Delete All Data", preferredStyle: .alert)
+            let alertAction = UIAlertAction(title: "Ok", style: .default) { (_) in
+                self.realmService.deleteData()
+            }
+            let alertCancel = UIAlertAction(title: "Cancel", style: .cancel) { (_) in }
+            alertController.addAction(alertAction)
+            alertController.addAction(alertCancel)
+            present(alertController, animated: true, completion: nil)
+        } else {
+            showMegaAlert("No Data to Delete!")
         }
-        let alertCancel = UIAlertAction(title: "Cancel", style: .cancel) { (_) in }
-        alertController.addAction(alertAction)
-        alertController.addAction(alertCancel)
-        present(alertController, animated: true, completion: nil)
     }
 }
